Added check user ID, Delete card, find 1 card

+++ controllers/cards.js
@@ -2,6 +2,7 @@ const Card = require('../models/card');
 const cookieParser = require('cookie-parser');
 
 module.exports.createCard = (req, res) => {
+  console.log(req.user._id);
   const { name, link } = req.body;
   const ownerId = req.user._id;
   console.log(req.body);
@@ -12,15 +13,34 @@ module.exports.createCard = (req, res) => {
 
 module.exports.deleteCard = (req, res) => {
   const { cardId } = req.params;
-  Card.findByIdAndRemove(cardId)
+  const ownerId = req.user._id;
+  Card.findById(cardId)
+    .then((card) => {
+      if (card.owner != ownerId) {
+        return res.status(401).send({ message: 'Вы не имеете доступ к удалению чужих карточек' })
+      }
+      Card.findByIdAndRemove(cardId)
+        .then((card) => res.send({ data: card }))
+        .catch(() => res.status(500).send({ message: 'Произошла ошибка' }));
+    })
+  // Card.findByIdAndRemove(cardId)
+  //   .then((card) => res.send({ data: card }))
+  //   .catch(() => res.status(500).send({ message: 'Произошла ошибка' }));
+};
+
+module.exports.getAllCards = (req, res) => {
+  Card.find({})
     .then((card) => res.send({ data: card }))
     .catch(() => res.status(500).send({ message: 'Произошла ошибка' }));
 };
 
-module.exports.getAllCards = (req, res) => {
-  console.log(req.cookies.jwt);
-};
-
+module.exports.getCard = (req, res) => {
+  const { cardId } = req.params;
+  console.log(cardId)
+  Card.findById(cardId)
+    .then((card) => res.send(card.owner))
+    .catch(() => res.status(500).send({ message: 'Произошла ошибка' }));
+}
 module.exports.likeCard = (req, res) => {
   const { cardId } = req.params;
   Card.findByIdAndUpdate(cardId, { $addToSet: { likes: req.user._id } }, { new: true })
